requirements.txt
Flask>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.1
openai>=1.40.0
